requirements.txt
einops==0.8.0
gradio==5.12.0
kornia==0.7.2
lpips==0.1.4
numpy==2.0.0
opencv-python==4.11.0.86
pandas==2.2.2
pillow==10.3.0
ptflops==0.7.3
pyiqa==0.1.13
pytorch-msssim==1.0.0
PyWavelets==1.8.0
PyYAML==6.0.1
scikit-image==0.24.0
scipy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.4